requirement.txt
accelerate==1.4.0
appnope==0.1.4
asttokens==2.4.1
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
dotenv==0.9.9
executing==2.0.1
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.29.2
idna==3.10
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.5
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.0
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==75.8.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
